Updating Trackable::association_hash to always write child relationship name.
Currently there are cases (on destroy mainly) where an embedded document
would write it's class name to the association hash name field, when it
fails to lookup the relationship meta information pertaining to it's parent.

Then, if you are trying to use that history track later on, it would
fail to be able to lookup it's trackable_parent, among other methods,
because it would try to traverse a relationship titled "Comment" instead
of a relationship titled "comments".

This fix makes it so a child document never fails to lookup the name of
the relation it's parent uses to refer to it. Thus class names never get
used in child documents.